MovieList added and Movies with list id

+++ JavaEE-Files/src/main/java/com/qa/persistence/repository/AccountRepository.java
@@ -13,9 +13,9 @@ public interface AccountRepository {
 
 	String createAccount(String account);
 
-	String deleteAccount(int accountId) throws AccountNotFoundException;
+	String deleteAccount(Long accountId) throws AccountNotFoundException;
 
-	String updateAccount(int accountId, String account) throws AccountNotFoundException;
+	String updateAccount(Long accountId, String account) throws AccountNotFoundException;
 	
 	String login(String account);
